requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
altgraph==0.17.4
anyio==4.4.0
apache-airflow==2.9.2
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.20.0
apache-airflow-providers-fab==1.2.0
apache-airflow-providers-ftp==3.10.0
apache-airflow-providers-http==4.12.0
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-microsoft-mssql==3.7.2
apache-airflow-providers-odbc==4.8.1
apache-airflow-providers-postgres==5.13.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
Automat==22.10.0
Babel==2.15.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
comm==0.1.4
config==0.5.1
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
contourpy==1.1.1
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.7
Django==4.2.6
dnspython==2.6.1
docutils==0.21.2
email_validator==2.2.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.4
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.42.1
frozenlist==1.4.1
fsspec==2024.6.1
google==3.0.0
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.23.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-re2==1.1.20240702
googleapis-common-protos==1.63.2
greenlet==3.0.1
grpcio==1.64.1
gspread==5.11.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
incremental==22.10.0
inflection==0.5.1
ipykernel==6.25.2
ipython==8.15.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
klib==1.1.2
lazy-object-proxy==1.10.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.1.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
mlxtend==0.23.0
more-itertools==10.3.0
multidict==6.0.5
nest-asyncio==1.5.7
numpy==2.0.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
outcome==1.2.0
packaging==24.1
pandas==2.2.2
paramiko==3.4.0
parsel==1.8.1
parso==0.8.3
pathspec==0.12.1
pefile==2023.2.7
pendulum==3.0.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.10.0
playwright==1.40.0
plotly==5.17.0
pluggy==1.5.0
prison==0.2.1
prompt-toolkit==3.0.39
Protego==0.3.0
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.5
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==11.0.1
Pygments==2.16.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.8.0
pymssql==2.3.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypyodbc==1.3.6
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.35.1
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.18.1
rsa==4.9
schedule==1.2.1
scikit-learn==1.3.1
scipy==1.11.2
Scrapy==2.11.0
screeninfo==0.8.1
seaborn==0.12.2
selenium==4.14.0
service-identity==23.1.0
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.2
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
time-machine==2.14.2
tldextract==5.1.1
tornado==6.3.3
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.11.1
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2024.1
uc-micro-py==1.0.3
undetected-chromedriver==3.5.5
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
w3lib==2.1.2
wcwidth==0.2.6
websockets==14.1
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
zope.interface==6.1
